Rename `models` package to `contracts`
